+++ x/evm/statedb/statedb.go
@@ -21,6 +21,9 @@ import (
 	"sort"
 
 	errorsmod "cosmossdk.io/errors"
+	sdkmath "cosmossdk.io/math"
+	"cosmossdk.io/store/cachemulti"
+	"cosmossdk.io/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
@@ -28,12 +31,20 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 )
 
+const StateDBContextKey = "statedb"
+
+type EventConverter = func(sdk.Event) (*ethtypes.Log, error)
+
 // revision is the identifier of a version of state.
 // it consists of an auto-increment id and a journal index.
 // it's safer to use than using journal index alone.
 type revision struct {
 	id           int
 	journalIndex int
+}
+
+func Transfer(db vm.StateDB, sender, recipient common.Address, amount *big.Int) {
+	db.(*StateDB).Transfer(sender, recipient, amount)
 }
 
 var _ vm.StateDB = &StateDB{}
@@ -45,7 +56,18 @@ var _ vm.StateDB = &StateDB{}
 // * Accounts
 type StateDB struct {
 	keeper Keeper
-	ctx    sdk.Context
+	// origCtx is the context passed in by the caller
+	origCtx sdk.Context
+	// ctx is a branched context on top of the caller context
+	ctx sdk.Context
+	// cacheMS caches the `ctx.MultiStore()` to avoid type assertions all the time, `ctx.MultiStore()` is not modified during the whole time,
+	// which is evident by `ctx.WithMultiStore` is not called after statedb constructed.
+	cacheMS types.CacheMultiStore
+
+	// the action to commit native state, there are two cases:
+	// if the parent store is not `cachemulti.Store`, we create a new one, and call `Write` to commit, this could only happen in unit tests.
+	// if the parent store is already a `cachemulti.Store`, we branch it and call `Restore` to commit.
+	commitMS []func()
 
 	// Journal of state modifications. This is the backbone of
 	// Snapshot and RevertToSnapshot.
@@ -65,19 +87,55 @@ type StateDB struct {
 
 	// Per-transaction access list
 	accessList *accessList
+
+	// events emitted by native action
+	nativeEvents sdk.Events
+
+	// handle balances natively
+	evmDenom string
+	err      error
 }
 
 // New creates a new state from a given trie.
 func New(ctx sdk.Context, keeper Keeper, txConfig TxConfig) *StateDB {
-	return &StateDB{
+	return NewWithParams(ctx, keeper, txConfig, keeper.GetParams(ctx).EvmDenom)
+}
+
+func NewWithParams(ctx sdk.Context, keeper Keeper, txConfig TxConfig, evmDenom string) *StateDB {
+	var (
+		cacheMS  types.CacheMultiStore
+		commitMS []func()
+	)
+	if parentCacheMS, ok := ctx.MultiStore().(cachemulti.Store); ok {
+		cacheMS = parentCacheMS.CacheMultiStore()
+		commitMS = append(commitMS, cacheMS.Write)
+	} else {
+		// in unit test, it could be run with a uncached multistore
+		if cacheMS, ok = ctx.MultiStore().CacheWrap().(cachemulti.Store); !ok {
+			panic("expect the CacheWrap result to be cachemulti.Store")
+		}
+		commitMS = append(commitMS, cacheMS.Write)
+	}
+	db := &StateDB{
+		origCtx:      ctx,
 		keeper:       keeper,
-		ctx:          ctx,
+		cacheMS:      cacheMS,
+		commitMS:     commitMS,
 		stateObjects: make(map[common.Address]*stateObject),
 		journal:      newJournal(),
 		accessList:   newAccessList(),
 
 		txConfig: txConfig,
+
+		nativeEvents: sdk.Events{},
+		evmDenom:     evmDenom,
 	}
+	db.ctx = ctx.WithValue(StateDBContextKey, db)
+	return db
+}
+
+func (s *StateDB) NativeEvents() sdk.Events {
+	return s.nativeEvents
 }
 
 // Keeper returns the underlying `Keeper`
@@ -127,16 +185,15 @@ func (s *StateDB) Exist(addr common.Address) bool {
 // or empty according to the EIP161 specification (balance = nonce = code = 0)
 func (s *StateDB) Empty(addr common.Address) bool {
 	so := s.getStateObject(addr)
-	return so == nil || so.empty()
+	if so == nil {
+		return true
+	}
+	return so.empty() && s.GetBalance(addr).Sign() == 0
 }
 
 // GetBalance retrieves the balance from the given address or 0 if object not found
 func (s *StateDB) GetBalance(addr common.Address) *big.Int {
-	stateObject := s.getStateObject(addr)
-	if stateObject != nil {
-		return stateObject.Balance()
-	}
-	return common.Big0
+	return s.keeper.GetBalance(s.ctx, sdk.AccAddress(addr.Bytes()), s.evmDenom)
 }
 
 // GetNonce returns the nonce of account, 0 if not exists.
@@ -227,7 +284,7 @@ func (s *StateDB) getStateObject(addr common.Address) *stateObject {
 		return nil
 	}
 	// Insert into the live set
-	obj := newObject(s, addr, *account)
+	obj := newObject(s, addr, account)
 	s.setStateObject(obj)
 	return obj
 }
@@ -236,27 +293,24 @@ func (s *StateDB) getStateObject(addr common.Address) *stateObject {
 func (s *StateDB) getOrNewStateObject(addr common.Address) *stateObject {
 	stateObject := s.getStateObject(addr)
 	if stateObject == nil {
-		stateObject, _ = s.createObject(addr)
+		stateObject = s.createObject(addr)
 	}
 	return stateObject
 }
 
 // createObject creates a new state object. If there is an existing account with
 // the given address, it is overwritten and returned as the second return value.
-func (s *StateDB) createObject(addr common.Address) (newobj, prev *stateObject) {
-	prev = s.getStateObject(addr)
+func (s *StateDB) createObject(addr common.Address) *stateObject {
+	prev := s.getStateObject(addr)
 
-	newobj = newObject(s, addr, Account{})
+	newobj := newObject(s, addr, nil)
 	if prev == nil {
 		s.journal.append(createObjectChange{account: &addr})
 	} else {
 		s.journal.append(resetObjectChange{prev: prev})
 	}
 	s.setStateObject(newobj)
-	if prev != nil {
-		return newobj, prev
-	}
-	return newobj, nil
+	return newobj
 }
 
 // CreateAccount explicitly creates a state object. If a state object with the address
@@ -270,10 +324,7 @@ func (s *StateDB) createObject(addr common.Address) (newobj, prev *stateObject) 
 //
 // Carrying over the balance ensures that Ether doesn't disappear.
 func (s *StateDB) CreateAccount(addr common.Address) {
-	newObj, prev := s.createObject(addr)
-	if prev != nil {
-		newObj.setBalance(prev.account.Balance)
-	}
+	s.createObject(addr)
 }
 
 // ForEachStorage iterate the contract storage, the iteration order is not defined.
@@ -298,23 +349,110 @@ func (s *StateDB) setStateObject(object *stateObject) {
 	s.stateObjects[object.Address()] = object
 }
 
+// TODO(zsystm): Use or remove this function, currently testing other way.
+//
+//	func (s *StateDB) snapshotNativeState() types.CacheMultiStore {
+//		snapshot := s.cacheMS.CacheMultiStore()
+//		s.commitMS = append(s.commitMS, snapshot.Write)
+//		return snapshot
+//	}
+func (s *StateDB) revertNativeStateToSnapshot(ms types.MultiStore) {
+	s.ctx = s.ctx.WithMultiStore(ms)
+	s.commitMS = s.commitMS[:len(s.commitMS)-1]
+}
+
+// ExecuteNativeAction executes native action in isolate,
+// the writes will be revert when either the native action itself fail
+// or the wrapping message call reverted.
+func (s *StateDB) ExecuteNativeAction(contract common.Address, converter EventConverter, action func(ctx sdk.Context) error) error {
+	orig := s.cacheMS
+	// cache the orig to avoid the cacheMS being modified by the action
+	cached := orig.CacheMultiStore()
+	s.commitMS = append(s.commitMS, cached.Write)
+	eventManager := sdk.NewEventManager()
+
+	s.ctx = s.ctx.WithMultiStore(cached)
+	if err := action(s.ctx.WithEventManager(eventManager)); err != nil {
+		// revert the state to the orig state
+		s.revertNativeStateToSnapshot(orig)
+		return err
+	}
+	events := eventManager.Events()
+	s.emitNativeEvents(contract, converter, events)
+	s.nativeEvents = s.nativeEvents.AppendEvents(events)
+	// journal for reverting case
+	s.journal.append(nativeChange{snapshot: orig, events: len(events)})
+	// update latest cacheMS to the new one
+	s.cacheMS = cached
+	return nil
+}
+
+// Context returns the current context for query native state in precompiles.
+func (s *StateDB) Context() sdk.Context {
+	return s.ctx
+}
+
 /*
  * SETTERS
  */
 
+// Transfer from one account to another
+func (s *StateDB) Transfer(sender, recipient common.Address, amount *big.Int) {
+	if amount.Sign() == 0 {
+		return
+	}
+	if amount.Sign() < 0 {
+		panic("negative amount")
+	}
+
+	coins := sdk.NewCoins(sdk.NewCoin(s.evmDenom, sdkmath.NewIntFromBigIntMut(amount)))
+	senderAddr := sdk.AccAddress(sender.Bytes())
+	recipientAddr := sdk.AccAddress(recipient.Bytes())
+	if err := s.ExecuteNativeAction(common.Address{}, nil, func(ctx sdk.Context) error {
+		return s.keeper.Transfer(ctx, senderAddr, recipientAddr, coins)
+	}); err != nil {
+		s.err = err
+	}
+}
+
 // AddBalance adds amount to the account associated with addr.
 func (s *StateDB) AddBalance(addr common.Address, amount *big.Int) {
-	stateObject := s.getOrNewStateObject(addr)
-	if stateObject != nil {
-		stateObject.AddBalance(amount)
+	if amount.Sign() == 0 {
+		return
+	}
+	if amount.Sign() < 0 {
+		panic("negative amount")
+	}
+	coins := sdk.Coins{sdk.NewCoin(s.evmDenom, sdkmath.NewIntFromBigInt(amount))}
+	if err := s.ExecuteNativeAction(common.Address{}, nil, func(ctx sdk.Context) error {
+		return s.keeper.AddBalance(ctx, sdk.AccAddress(addr.Bytes()), coins)
+	}); err != nil {
+		s.err = err
 	}
 }
 
 // SubBalance subtracts amount from the account associated with addr.
 func (s *StateDB) SubBalance(addr common.Address, amount *big.Int) {
-	stateObject := s.getOrNewStateObject(addr)
-	if stateObject != nil {
-		stateObject.SubBalance(amount)
+	if amount.Sign() == 0 {
+		return
+	}
+	if amount.Sign() < 0 {
+		panic("negative amount")
+	}
+	coins := sdk.Coins{sdk.NewCoin(s.evmDenom, sdkmath.NewIntFromBigInt(amount))}
+	if err := s.ExecuteNativeAction(common.Address{}, nil, func(ctx sdk.Context) error {
+		return s.keeper.SubBalance(ctx, sdk.AccAddress(addr.Bytes()), coins)
+	}); err != nil {
+		s.err = err
+	}
+}
+
+// SetBalance is called by state override
+func (s *StateDB) SetBalance(addr common.Address, amount *big.Int) {
+	if err := s.ExecuteNativeAction(common.Address{}, nil, func(ctx sdk.Context) error {
+		return s.keeper.SetBalance(ctx, addr, amount, s.evmDenom)
+	}); err != nil {
+		s.err = err
 	}
 }
 
@@ -336,10 +474,16 @@ func (s *StateDB) SetCode(addr common.Address, code []byte) {
 
 // SetState sets the contract state.
 func (s *StateDB) SetState(addr common.Address, key, value common.Hash) {
-	stateObject := s.getOrNewStateObject(addr)
-	if stateObject != nil {
-		stateObject.SetState(key, value)
-	}
+	stateObject := s.getOrNewStateObject(addr) // TODO(zsystm): nil check?
+	stateObject.SetState(key, value)
+}
+
+// SetStorage replaces the entire storage for the specified account with given
+// storage. This function should only be used for debugging and the mutations
+// must be discarded afterwards.
+func (s *StateDB) SetStorage(addr common.Address, storage Storage) {
+	stateObject := s.getOrNewStateObject(addr) // TODO(zsystm): nil check?
+	stateObject.SetStorage(storage)
 }
 
 // Suicide marks the given account as suicided.
@@ -353,38 +497,55 @@ func (s *StateDB) Suicide(addr common.Address) bool {
 		return false
 	}
 	s.journal.append(suicideChange{
-		account:     &addr,
-		prev:        stateObject.suicided,
-		prevbalance: new(big.Int).Set(stateObject.Balance()),
+		account: &addr,
+		prev:    stateObject.suicided,
 	})
 	stateObject.markSuicided()
-	stateObject.account.Balance = new(big.Int)
+
+	// clear balance
+	balance := s.GetBalance(addr)
+	if balance.Sign() > 0 {
+		s.SubBalance(addr, balance)
+	}
 
 	return true
 }
 
-// PrepareAccessList handles the preparatory steps for executing a state transition with
-// regards to both EIP-2929 and EIP-2930:
+// Prepare handles the preparatory steps for executing a state transition with.
+// This method must be invoked before state transition.
 //
+// Berlin fork:
 // - Add sender to access list (2929)
 // - Add destination to access list (2929)
 // - Add precompiles to access list (2929)
 // - Add the contents of the optional tx access list (2930)
 //
-// This method should only be called if Yolov3/Berlin/2929+2930 is applicable at the current number.
-func (s *StateDB) PrepareAccessList(sender common.Address, dst *common.Address, precompiles []common.Address, list ethtypes.AccessList) {
-	s.AddAddressToAccessList(sender)
+// Potential EIPs:
+// - Reset access list (Berlin)
+// - Add coinbase to access list (EIP-3651)
+// - Reset transient storage (EIP-1153)
+func (s *StateDB) PrepareAccessList(
+	sender common.Address,
+	dst *common.Address,
+	precompiles []common.Address,
+	list ethtypes.AccessList,
+) {
+	// Clear out any leftover from previous executions
+	al := newAccessList()
+	s.accessList = al
+
+	al.AddAddress(sender)
 	if dst != nil {
-		s.AddAddressToAccessList(*dst)
+		al.AddAddress(*dst)
 		// If it's a create-tx, the destination will be added inside evm.create
 	}
 	for _, addr := range precompiles {
-		s.AddAddressToAccessList(addr)
+		al.AddAddress(addr)
 	}
 	for _, el := range list {
-		s.AddAddressToAccessList(el.Address)
+		al.AddAddress(el.Address)
 		for _, key := range el.StorageKeys {
-			s.AddSlotToAccessList(el.Address, key)
+			al.AddSlot(el.Address, key)
 		}
 	}
 }
@@ -451,18 +612,39 @@ func (s *StateDB) RevertToSnapshot(revid int) {
 // Commit writes the dirty states to keeper
 // the StateDB object should be discarded after committed.
 func (s *StateDB) Commit() error {
+	// if there's any errors during the execution, abort
+	if s.err != nil {
+		return s.err
+	}
+
+	// commit the native cache store first,
+	// the states managed by precompiles and the other part of StateDB must not overlap.
+	// after this, should only use the `origCtx`.
+	ei := len(s.commitMS) - 1
+	for si := ei; si >= 0; si-- {
+		// commit the native state
+		// child store should be committed first
+		s.commitMS[si]()
+	}
+	if len(s.nativeEvents) > 0 {
+		s.origCtx.EventManager().EmitEvents(s.nativeEvents)
+	}
+
 	for _, addr := range s.journal.sortedDirties() {
 		obj := s.stateObjects[addr]
 		if obj.suicided {
-			if err := s.keeper.DeleteAccount(s.ctx, obj.Address()); err != nil {
+			if err := s.keeper.DeleteAccount(s.origCtx, obj.Address()); err != nil {
 				return errorsmod.Wrap(err, "failed to delete account")
 			}
 		} else {
-			if obj.code != nil && obj.dirtyCode {
-				s.keeper.SetCode(s.ctx, obj.CodeHash(), obj.code)
+			codeDirty := obj.codeDirty()
+			if codeDirty && obj.code != nil {
+				s.keeper.SetCode(s.origCtx, obj.CodeHash(), obj.code)
 			}
-			if err := s.keeper.SetAccount(s.ctx, obj.Address(), obj.account); err != nil {
-				return errorsmod.Wrap(err, "failed to set account")
+			if codeDirty || obj.nonceDirty() {
+				if err := s.keeper.SetAccount(s.origCtx, obj.Address(), obj.account); err != nil {
+					return errorsmod.Wrap(err, "failed to set account")
+				}
 			}
 			for _, key := range obj.dirtyStorage.SortedKeys() {
 				value := obj.dirtyStorage[key]
@@ -470,9 +652,33 @@ func (s *StateDB) Commit() error {
 				if value == obj.originStorage[key] {
 					continue
 				}
-				s.keeper.SetState(s.ctx, obj.Address(), key, value.Bytes())
+				s.keeper.SetState(s.origCtx, obj.Address(), key, value.Bytes())
 			}
 		}
 	}
 	return nil
+}
+
+func (s *StateDB) emitNativeEvents(contract common.Address, converter EventConverter, events []sdk.Event) {
+	if converter == nil {
+		return
+	}
+
+	if len(events) == 0 {
+		return
+	}
+
+	for _, event := range events {
+		log, err := converter(event)
+		if err != nil {
+			s.ctx.Logger().Error("failed to convert event", "err", err)
+			continue
+		}
+		if log == nil {
+			continue
+		}
+
+		log.Address = contract
+		s.AddLog(log)
+	}
 }